Fixed error 500
An error would happen 'Failed to load API definition'. To fix this I had to go to the Developer Tools in the browser. Then go to console and click a some link with json. It showed a stack  trace and the error was that Swagger confused a public method of FlightController for a Http endpoint which it wasnt tagged as. I had to make the method internal to solve this.